+++ NewSpellsAPI/Sigils/GiveStats.cs
@@ -16,7 +16,7 @@ namespace Infiniscryption.Spells.Sigils
         public override bool RespondsToSacrifice() => true;
         public override bool RespondsToSlotTargetedForAttack(CardSlot slot, PlayableCard attacker)
         {
-            if (slot.Card != null)
+            if (base.Card.Info.IsSpell() && slot.Card != null)
                 return base.Card.OpponentCard == slot.Card.OpponentCard;
 
             return false;